Refactor screensaver config, channel video config, and loading screen config
They now all use the new config.json file, and their config formats have been streamlined.

Also updates docs on encoding h264 video in real time.

Also switches ffprobe to output in json instead of csv, because the csv field order output was not necessarily the same as the order in which you declared the fields you wanted to be output. Thus json will be easier to parse as it is not dependent on order.


Former-commit-id: d1e3739e73260d5d1d9e320d3df6a23de01dd37b