refactor how audio muting works
Mute audio in the receiver rather than in the broadcaster. This is because loading screen videos are played 100% locally by the receivers -- the broadcaster does not transmit the loading screen videos to the receivers. So if muting is done in the broadcaster, we will be unable to mute the loading screen videos.